+++ src/mkvtoolnix-gui/header_editor/time_value_page.cpp
@@ -1,11 +1,12 @@
 #include "common/common_pch.h"
 
 #include <QDateTimeEdit>
-#include <QTimeZone>
 
 #include "common/qt.h"
 #include "mkvtoolnix-gui/header_editor/time_value_page.h"
-#include "mkvtoolnix-gui/util/string.h"
+#include "mkvtoolnix-gui/main_window/main_window.h"
+#include "mkvtoolnix-gui/util/date_time.h"
+#include "mkvtoolnix-gui/util/settings.h"
 
 namespace mtx::gui::HeaderEditor {
 
@@ -19,6 +20,7 @@ TimeValuePage::TimeValuePage(Tab &parent,
                              translatable_string_c const &description)
   : ValuePage{parent, topLevelPage, master, callbacks, ValueType::String, title, description}
 {
+  connect(MainWindow::get(), &MainWindow::preferencesChanged, this, &TimeValuePage::showInRequestedTimeSpec);
 }
 
 TimeValuePage::~TimeValuePage() {
@@ -26,12 +28,15 @@ TimeValuePage::~TimeValuePage() {
 
 QWidget *
 TimeValuePage::createInputControl() {
+  auto &cfg = Util::Settings::get();
+
   if (m_element)
-    m_originalValue = QDateTime::fromMSecsSinceEpoch(static_cast<EbmlDate *>(m_element)->GetEpochDate() * 1000, Qt::UTC).toLocalTime();
+    m_originalValueUTC = QDateTime::fromMSecsSinceEpoch(static_cast<EbmlDate *>(m_element)->GetEpochDate() * 1000, Qt::UTC);
 
   m_dteValue = new QDateTimeEdit{this};
   m_dteValue->setCalendarPopup(true);
-  m_dteValue->setDateTime(m_originalValue);
+  m_dteValue->setTimeSpec(cfg.m_headerEditorDateTimeInUTC ? Qt::UTC            : Qt::LocalTime);
+  m_dteValue->setDateTime(cfg.m_headerEditorDateTimeInUTC ? m_originalValueUTC : m_originalValueUTC.toLocalTime());
   m_dteValue->setDisplayFormat(Q("yyyy-MM-dd hh:mm:ss"));
 
   return m_dteValue;
@@ -40,7 +45,9 @@ TimeValuePage::createInputControl() {
 QString
 TimeValuePage::originalValueAsString()
   const {
-  return Util::displayableDate(m_originalValue);
+  auto &cfg = Util::Settings::get();
+
+  return Util::displayableDate(cfg.m_headerEditorDateTimeInUTC ? m_originalValueUTC : m_originalValueUTC.toLocalTime());
 }
 
 QString
@@ -51,7 +58,9 @@ TimeValuePage::currentValueAsString()
 
 void
 TimeValuePage::resetValue() {
-  m_dteValue->setDateTime(m_originalValue);
+  auto &cfg = Util::Settings::get();
+
+  m_dteValue->setDateTime(cfg.m_headerEditorDateTimeInUTC ? m_originalValueUTC : m_originalValueUTC.toLocalTime());
 }
 
 bool
@@ -63,6 +72,24 @@ TimeValuePage::validateValue()
 void
 TimeValuePage::copyValueToElement() {
   static_cast<EbmlDate *>(m_element)->SetEpochDate(m_dteValue->dateTime().toUTC().toMSecsSinceEpoch() / 1000);
+}
+
+void
+TimeValuePage::showInRequestedTimeSpec() {
+  auto &cfg    = Util::Settings::get();
+  auto current = m_dteValue->dateTime();
+
+  m_dteValue->setTimeSpec(cfg.m_headerEditorDateTimeInUTC ? Qt::UTC : Qt::LocalTime);
+  m_dteValue->setDateTime(cfg.m_headerEditorDateTimeInUTC ? current.toUTC() : current.toLocalTime());
+}
+
+QString
+TimeValuePage::note()
+  const {
+  auto &cfg = Util::Settings::get();
+  auto now  = QDateTime::currentDateTime();
+
+  return cfg.m_headerEditorDateTimeInUTC || !now.offsetFromUtc() ? QY("The date & time shown is in UTC.") : QY("The date & time shown is in your local time zone.");
 }
 
 }