+++ src/mkvtoolnix-gui/header_editor/time_value_page.h
@@ -15,7 +15,7 @@ class Tab;
 class TimeValuePage: public ValuePage {
 public:
   QDateTimeEdit *m_dteValue{};
-  QDateTime m_originalValue;
+  QDateTime m_originalValueUTC;
 
 public:
   TimeValuePage(Tab &parent, PageBase &topLevelPage, EbmlMaster &master, EbmlCallbacks const &callbacks, translatable_string_c const &title, translatable_string_c const &description);
@@ -24,9 +24,11 @@ public:
   virtual QWidget *createInputControl() override;
   virtual QString originalValueAsString() const override;
   virtual QString currentValueAsString() const override;
+  virtual QString note() const override;
   virtual void resetValue() override;
   virtual bool validateValue() const override;
   virtual void copyValueToElement() override;
+  virtual void showInRequestedTimeSpec();
 };
 
 }